DBLD: Update package list files
  - new helpers/packages.manifest: contains packages for apt
and yum based package managers. Each package can belongs to a
distro group (centos or debian or ubuntu), or to an exact OS
(ubuntu-trusty) group. During the installation package will
be selected only if there is an exact match for an OS group OR
the current distroname matches
  - new helpers/pip_packages.manifest: works as same mechanism as
packages.manifest
  - install_apt_packages, install_yum_packages, install_pip_packages
are the new high level functions for package installations
  - old packages list files was removed
  - old package install mechanism (native, OBS, Epel) was
removed from dockerfiles
  - dbld/rules apt-[image] target is removed


Signed-off-by: Andras Mitzki <andras.mitzki@balabit.com>
Signed-off-by: Laszlo Szemere <laszlo.szemere@balabit.com>