logthrdestdrv: register stats in _start_workers()
Instead of registering stat counters in _dd_init(), we do that in
start_workers as some of the initialization for that might happen in the
derived class.

For instance, our java binding calls into java to format the stat counter
key. This preparation happens _after_
log_threaded_dest_driver_init_instance(), thus we can't possibly do the
stats registration.

_start_workers() on the other hand is called _after_ all initialization
happened and workers are ready to start. By moving the stats registration
there we give derived classes a chance to set up everything.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>