refactor: eliminating (reliable flag) if inside qdisk_init
There is a branching inside qdisk_init based on type information, that
I eliminate in this patch.

LogQueueDiskReliable and LogQueueDiskNonReliable both are inherited
from LogQueueDisk, which contains a QDisk instance. qdisk_init would
be responsible for initializing the QDisk instance. Currently, as
qdisk_init contains a branching based on reliable and non-reliable,
which is wrong way of dependency.

As qdisk_init should not know what is the magic id, its parameter set
is extended with it (called file_id in the code).

Moreover, mem_buf_size is reliable diskqueue specific, so it should
not be set inside qdisk_init.

The magic id and mem_buf_length cannot be changed once the queues
are created. So these assignments can go into the respective new methods.

There is one problem though. When an error occurs during reading from
the disk queue, the diskqueue needs to be restarted. During that
qdisk_init was called again. However, the restart function was defined
in logqueue-disk, which is again unaware of the type of the underlying
queue, so a restart virtual function was necessary to be used in
LogQueueDisk, that needs to be overridden in LogQueueDiskNonReliable
and LogQueueDiskReliable. Luckily, there was already a restart method
unused, that I could use for the cause.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>