non-reliable diskq: fixing a false positive corruption detection
Previously, when a non-reliable diskqueue file is loaded, even though
the qout, qbacklog qoverflow queues were truncated from the file,
their pointers were not reset. As a consequence, the diskqueue file is
kept in an invalid state while syslog-ng was running. These pointers were
only set to valid values only when syslog-ng stopped gracefully, while
saving the diskqueue.

Suppose more messages are added to the queue, so that the write_head
passes the invalid qout_ofs, then crash syslog-ng. In that
case, the original invalid qout_ofs was kept in the diskqueue file,
which is detected as a corruption at the next start, though the
content of the disk queue was valid: at least the memory queue point
of view, as everything in the memory was lost.

This patch resets the queue pointers after load, fixing this scenario.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>