http: free workers in non-compat mode
In case of compat mode, the embedded structure is still added to
self->workers. However in that case, the embedded structure must not
be freed. So we need to free in a for loop only in non compat mode.

self->workers_started must be set for another reason: when syslog-ng
reloaded to an invalid configuration, syslog-ng will try to fall back
to the original config. In that case the init will be called two
times: self->workers array is recreated, but workers_started remains
the same. In the end, the workers would be put after self->workers,
resulting in an invalid write.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>
Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>
Signed-off-by: László Várady <laszlo.varady@balabit.com>