apphook: clear up apphook concepts to ease understanding
We seem to have two use-cases around apphooks, one is to tie some
functionality along an event happening in the process life-cycle (e.g.
run something once daemonized or run something before shutdown).

The other use-case can happen any number of times during the execution of
syslog-ng, e.g. to reopen all destination files upon SIGUSR1 or
when the configuration changes.

This patch separates these use-cases a bit more clearly in the source.  I've
introduced a "threshold" value in the apphook enum, anything below the
threshold will cause the state to be changed, anything greater that the
threshold will not.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>