mainloop: move control command implementations to a separate module
This patch extracts MainLoop specific control commands into a separate
module and turns control-commands.c into a generic container that manages
the command list.

This has a number of consequences and is still a step into the right
direction with some number of uglyness remaining (like the commands global
variable in control-commands).

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>