non-reliable-diskq: refactor - remove code duplication on queue load
Instaed of _try_to_load_(qoverflow,qout,qbacklog), an universal
_try_to_load_queue is created. For that sake, a LoadParameters struct
was needed that contains the offsets etc that is passed to those
functions. There was a problem with the msg_error about inconsistent
header: as it contained the queue type in the description, that should
come from function parameter. So the description could not have been
static string. To fix this, the error message is slightly changed: the
queue type is passed as an evt_tag.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>