csv-scanner: refactor to use an explicit state
This patch adds an explicit state variable to CSVScanner to track where
we are in processing, e.g. before even we started, processing columns or
at the very end.

This makes some conditions more obvious and the return values of various
functions more consistent.

NOTE: this patch is not without side effects, one intended change is
represented by a followup patch on csv-parser(), where empty messages will
not be accepted if drop-invalid is true.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>