lib/cfg-parser: in error reports, use arrow printout for all lines of the error
Instead of just marking up the first line (like we do with the underline),
the arrow on the left-hand side can point to all erroring lines, so
do that.

For instance:

[2018-10-26T16:13:25.874700] Mandatory parameter not overridden; parameter='foo', reference='etc/syslog-ng-splunk.conf:87:17'
Error parsing log statement, syntax error, unexpected LL_ERROR, expecting '}' in etc/syslog-ng-splunk.conf:
82      log {
83              source(s_local);
84      	#parser { app-parser(topic(syslog)); };
85      	filter { netmask("127.0.0.1/32"); tags('foobar'); };
86              destination {
87---->                 splunk-hec2(token("367ffe59-d1d8-484b-8508-94410fd119dd")
87---->                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
88----> 			   url("http://localhost:8080/services/collector/event")
89----> 			   workers(4)
90---->                            fields()
91----> 		);
92              };


Previously only the line with the underline would have used the arrows
on the left.