cfg-lex: fix location tracking for strings and blocks
strings, qstrings, block, block_string and block_qstring states were
greedy of newlines, so they were not reported properly in location
tracking, causing the error reports to be misleading.

This patch makes sure that newline characters in these states
are both added to the string value and the location is properly updated.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>