afsocket: add support for so-reuseport() option
This patch adds support for enabling SO_REUSEPORT on systems that support
it.  For a complete documentation on SO_REUSEPORT see the socket(7) manual
page.

The point is that when this is enabled, the kernel allows multiple UDP
sockets to be bound to the same port, and the kernel would load-balance
incoming UDP datagrams to the sockets.  The sockets are distributed based on
the hash of (srcip, dstip, srcport, dstport), so the same listener should be
receiving packets from the same endpoint.

	source {
		udp(so-reuseport(1) port(2000) persist-name("udp1"));
		udp(so-reuseport(1) port(2000) persist-name("udp2"));
		udp(so-reuseport(1) port(2000) persist-name("udp3"));
		udp(so-reuseport(1) port(2000) persist-name("udp4"));
	};

It is even possible to customize the load balance logic using BPF/eBPF
programs but for now the default behevior is perfectly fine.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>