http: flush when originator of the message being inserted is suspended
Before this patch http destination can be configured in such a way that
syslog-ng stop reading messages for flush-timeout
from the LogSource instance which is connected to a HTTP destination.
This can lead to performance issues (as users can set flush-timeout to
a high value).

Reason:
 * when iw_size <= worker-num * flush-lines then the source become
suspended before a batch would reach the flush-lines limit, so flush
won't be triggered until flush-timeout not elapsed (default: 10sec)

note: this problem can be also triggered with flush-bytes, but it is
easier to explain when using msg-count.

Some drawbacks:
 * when lots of sources are connected to a single http dst, then this
solution could result in unwanted calls of flush...
 * when a source is connected to two destinations(one of them is http
dst) then it is still possible that the source remain in blocked state
-> a default value for flush-timeout could resolve this

Signed-off-by: Laszlo Budai <laszlo.budai@balabit.com>