kafka: combine initial kafka and kafka-c modules
Kafka support was initially worked on independently in two independent
solutions, once was aimed at the creation of a kafka source, whereas the
destination was created separately.

This commit pulls in the destination into the same module that was used by
the source, but keeping the source specific pieces in followup patches as
the expectation is that the destination will be merged first.

Signed-off-by: Kokan <kokaipeter@gmail.com>