diskq: fix truncating problem of non-rel dq on load
On loading exiting disk-queues it was possible that a disk-queue only
contained saved memory queues (e.g. qout, qbacklog) and the disk-buf-size part was empty.
Due to the faulty logic of truncate on load, we had empty disk-queues with possibly big sizes.
Also, the truncating logic assumed that qout will be the lowest offset among the saved
memory queues, however it is possible that only qbacklog is saved to the disk, which
resulted into a huge empty disk-queue.

- file_size initialization fixed: prev. it was possible to have it 0
if qout_ofs was 0. It should be 4k size by default as file_size is following
write_head (see push_tail)