csv-scanner: add unit test program at this level
So far only csv-parser() had unit tests, but that's a layer a bit too high
and csv-parser() is now directly used by other callers
(e.g. add-contextual-data()), so it makes sense to write a unit test at this
level.

This is just an initial version, some of the coverage should be moved
here from csv-parser.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>