csv-scanner: move the "drop-invalid" flag and push it to csv-parser
The only case where the CSV input is not accepted when the input
contains more or less columns that was expected. With this flag,
csv-scanner was doing this decision but the only reason it used this flag
was to report whether it finished successfully.

This patch separates the two concerns, csv-scanner will be able to report
if it properly processed its entire input and that it fit into the
set of columns specified. Then, something like csv-parser() can use this
information to decide whether to drop the message or not.

Dropping a message is not a "concept" csv-scanner is capable of doing, so
splitting this up this way makes the code a bit cleaner.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>