csvparser: don't accept empty message if "drop-invalid" is set
The latest refactors in csv-scanner has one side-effect, empty input lines
will not be accepted by csv-parser(), which has been before the refactors
were applied.

This patch only adds a testcase to test_csvparser, but in reality the
previous refactor is causing it by more accurately tracking if we actually
finished or not.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>