xml-scanner: initialize key of scanner with argument
Scanner's key member can be initialized with a prefix.
With this option we can save some memory copy on every message setting
in the parser, though it's a bit extending the scanner's responsibilities.
(And we would save memory allocation parser side would allocate memory instead of using a buffer).

Other options is that we keep the prefix entirely on parser side,
though when we receive a name-value pair from the scanner we
have to concatenate the prefix and the given name before setting
the LogMessage.
And to be able to use the prefix in the push_function, we need
to pass it, thus not only the LogMessage, but the prefix
has to be included in the user_data.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>