logsource: using msg_diagnostics instead of trace to avoid inf. loop
When internal source is used and trace level enabled, some messages
generate infinite other messages. For example the triggering message
adds a new message to the internal message, which is reported in
log_source_queue, that will trigger a new message etc.

The infinite loop is broken by not generating messages into the
internal queue but written to stderr.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>