xml-scanner: Buffer partial text contents inside xml-scanner
Partial text contents now buffered inside xml-scanner
until the end tag is not found.
Previously, in case of mixed contents we pushed the partial text contents
separately:
```
<tag1>foo<tag2>noise</tag2>bar</tag1>
```
The above xml had resulted in `tag1 = foo` and `tag1 = bar` name-value pairs,
which got concatenated on the xml-parser side.
(it had to check for existing name-value pairs in LogMessage).

The problem is that for array-like xml structures this would look the same,
while it shouldn't as they are 2 separate name-value pairs.
```
<root>
<tag1>foo</tag1>
<tag1>bar</tag1>
</root>
```

If we buffer partial text contents, we can help the scanner-users (e.g. parser)
distinguish mixed contents and array-like elements.

The buffering implemented by stacking the current text content, and
whenever a start/end tag found we push/pop the current text content.
The push-key-value() call moved from text() to end() method so we don't
push chunks.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>