LogQueue: added logs_per_throttle_bucket
The LogQueue's throttle_buckets variable counts how many
logs can be popped from the queue, and its throttle
variable is responsible for calculating the right amount
of sleep, so we do not exceed the message / sec limit.

Currently these two variables are equal, so throttle
limits for log lines sent, and not batches sent,
which is okay, when we send 1 log line / batch,
but if we batch several log lines to 1 batch it
causes unexpected behavior.

In case of batching, the initial value of the LogQueue's
throttle_buckets should be batch_lines * throttle, so
the LogQueue can grab enough log lines for the batch,
but throttle should be intact, because the sleep is
calculated from it.

However LogQueue should not know about batch_lines,
but we can add an option to have a multiplicand relation
between throttle and throttle buckets, so the unchanged
throttle will cause the right amount of sleep, and the
multiplied throttle_buckets will provide the right amount
of log lines, which will be batched later in the Threaded
Destination's _perform_insert() method.

Signed-off-by: Attila Szakacs <attila.szakacs@balabit.com>