configure: fix "--with-docbook*" option usage
"--with-docbook-dir" option changed to "--with-docbook"

In dbld and in travis the "--with-docbook=" option is used, which does not exist,
therefore we always use the online version.

At first it seems trivial to change travis, dbld, etc to use the actual option,
although the config option name is misleading.
According to `xsltproc` man page, it is a file uses as an XML STYLESHEET.

Therefore I've changed the configure option and not the users of the option.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>