logthrdestdrv: batch_lines based flush is controlled by superclass
Drivers can still flush early (like http and flush_bytes).

I need this patch, because there is an inconvenient dependency: the
batching modules need to have access to logthrdestdrv, because they
need to ask if a batch is full or not, so that they know if they need
to flush. The problem is most visible in case of elastic search (see
next patch), when a completely separate module is responsible for
batch management, but it would still need to call logthrdestdrv
related functions.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>