str-timestamp: move syslog-timestamp parsing to /lib
This patch moves all of the syslog related timestamp parsing to the
core, together with the rest of the timestamp parsing functions.

This is actually also a license change, from GPL2+ to LGPL2+, due to
the different licensing strategies for the two components. This is OK,
as all of the authors in the respective module are/were Balabit employees:

$ git blame origin/master syslog-format.c | cut -c 50-67 | sort | uniq -c
     67 Antal Nemes
    920 Balazs Scheidler
     43 bkil-syslogng
     12 Gergely Nagy
    221 Juhász Viktor
      1 kokan
      7 MÓZES Ádám Istv
     43 Noémi Ványi
      3 Tibor Benke
      4 Viktor Juhasz

Also, this patch removes the macOS ld.so wizardry, as it has failed to
resolve the situation when both the code under test and the time caching
functions are in the same dylib.  And since this is the case once we move
the stuff in the core, I got rid off the ld.so stuff and used an alternative
solution.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>