control: delegate connection closing to ControlServer
It is the ControlServer class that creates ControlConnection instances,
so their destruction should also be part of their responsibility.

This patch does exactly that, instead ControlConnection commiting suicide,
it'll call control_server_connection_closed() to take care of it.

This also removes complicated I/O related calls from the
constructor/destructors that was spotted by @furiel.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>