timeutils: add WallClockTime class and implementation
This patch adds a simple wrapper over "struct tm" with fields that are
not portable across all platforms. For instance, tm_gmtoff is not
present on a number of platforms but would be very useful internally,
so we _always_ have the capability to represent a broken-down time with
all the properties needed for proper timezone reference.

The design of this class is so that we don't need to copy the contents
of "struct tm" around, rather we contain one instance and
provide wrapper macros so that fields that are present in struct tm are
used from there, those that aren't will be part of the wrapper structure.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>