logstamp: remove duplicate extremal value for timezone
An earlier patch introduced a second extremal "uninitialized" value for
the zone_offset field (just recently renamed to ut_gmtoff). This
is being used for the "PROCESSED" timestamp, so that during serialization
this timestamp is automatically replaced with the time of the
serialization, to avoid multiple, parallel changes to the same LogMessage
instance.

The extremal value chosen to detect this uninitialized state was G_MININT32,
however everywhere else in the code we were using -1.

It is true that -1 could be an actual GMT offset, however -1 is not
actually a timezone, as the smallest timezone difference is at
15 minute boundary. Also, -00:00 as a timezone is generally used to indicate
"unknown" timezone value in various RFCs (e.g RFC822 and derivatives).

So this nicely coincides with our choice of -1, which if not initialized
will be produced as -00:00, instead of +00:00 that we use when UTC is used.

Also, instead of using a constant to indicate uninitializedness, use the
already existing log_stamp_is_timezone_set() function.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>