timeutils: refactor ISO timestamp parsing to avoid using localtime()
Previously localtime() (more precisely its cached variant) was used
initialize the members of the struct tm instance where we parsed the bits
of the ISO timestamp into. This was used for portable initialization
of even unportable fields.

Since WallClockTime as a wrapper ensures that all fields are well known,
we can drop this behavior and the localtime call as well.

This follows what already happened in the BSD counterpart.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>