timeutils: split timestamp parsing and UnixTime conversion
This patch finally splits the code that parses a timestamp from the one
that takes a parsed timestamp and converts it to a time_t.

Makes parameter lists much simpler and the timestamp parsing functions
pretty similar to other timestamp related functions in a sense that
they return their results in a WallClockTime struct.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>