pytest_framework: timeout parameter for wait_until_true
Prior this patch, pytest_framework did not support specifying timeout
as parameter for `wait_until_true`. Instead, a global variable was
responsible for the timeout. The default timeout (10s) was good for
the functional tests, but not for the unit tests, which executed
properly but unnecessarily long. For unit tests this global variable
was changed to 0.1s, which is too small for the functional
tests. Consequently, unit tests and functional tests cannot not be
executed together.

This patch adds another `wait_until_true function`:
`wait_until_true_custom` which accepts timeout as parameter. For
simple cases the original can be used, however when needed, the
complex version is available.

The way of parameterization (arguments as tuple) was motivated from
the way `threading.Thread` works.
One inconvenient thing worth mentioning: in python - single element
tuples need explicit comma ... `(1) == 1`. One needs to use `(1,)`.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>