filter-re: ref/unref NVTable around regex eval
The regex evaluation can change the message itself, by adding
values to its NVTable.
The change can result in reallocating the message's NVTable, if
there is insufficient space for the newly added value.
We do not want this to happen, because we are pointing to a value
of it.

When we reference the NVTable, if it needs more space, it will be
copied, instead of being moved. After we are done using it, we can
unref it, which will result in freeing it, if it was copied during
the evaluation, so memory will not be leaked.

Signed-off-by: Attila Szakacs <attila.szakacs@balabit.com>