pytest_framework: moving version setting into syslog-ng temporarily
The goal is to be able to create SyslogNgConfig independently from
SyslogNg. For that I temporarily move the version setting from the
constructor to SyslogNg.start(), until I eliminate the __instances
variable. By the end of the patchet, I will move the version string
back to the constructor of SyslogNgConfig.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>