pytest_framework: moving version setting back to the constructor
This patch reverts/reimplements the original api that was changed in
the beginning of this patchset: the config version is initialized in
the SyslogNgConfig constructor. TestCase.new_config will receive the
config version through a session fixture. As syslog-ng binary is
executed in the fixture, it is worth setting the scope to session, so
that the binary is executed once, and not in every testcase. To avoid
scope conflict, I needed to change the scope of installdir too.