threaded-dest: make persist_name local
Before this commit, the worker's queue's `persist_name`
was stored in a static string, and it was accessed and
changed by multiple threads in the same time.

The static string was there to avoid having to care about
malloc/free, it is not performance critical, so it can be
changed to local.

Signed-off-by: Attila Szakacs <attila.szakacs@balabit.com>