value-pairs: add an --omit-empty-values option
This new option will allow a template function to skip name-value pairs
that are empty. This used to be the default behavior until unset() was
introduced, since which time we only omit unset name-value pairs.

For example, ElasticSearch needs a transaction description as a JSON
before each record, that can be generated like this:

$(format-json --omit-empty-values index._type=`type` index._index=`index` index._id=`custom_id`)

If $(format-json) would include index._id as an empty string, that would
invalidate the JSON in the eyes of ElasticSearch.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>