msg-format: move LM_TS_STAMP initialization from logsource if keep_timestamp() is false
Originally, keep-timestamp(no) was implemented by logsource, e.g.  any
source which posted a message gets it LM_TS_STAMP overwritten in case
keep-timestamp() is unset.

There's one problem however, in case the source has an explicit time-zone()
setting, we would like the so initialized timestamp in that timezone.

This means that setting LM_TS_STAMP to LM_TS_RECVD should have at least one
local, source specific setting.  An option which is not available at the
LogSource layer.

This patch moves that to log_msg_new(), after the point where MsgFormat got
a chance to parse the message.  We still have recv-time-zone() here and we
are certain that we preceed the pipeline processing.

I don't really like how MsgFormat and LogMessage is deeply coupled, but that
will be another time, when I break that coupling.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>