Read acknowledgements send by Riemann
When Riemann receive a message, it sends an acknowledgements to the
sending process when the connexion is over TCP (TCP or TLS mode of
operation).  Not reading this acknowledgement make these messages
fill-in the Recv-Q on the system running syslog-ng.

The riemann_communicate() function of riemann-c-client is a wrapper
around riemann_client_send_message_oneshot() that also reads this
acknowledgement if it exist, or mocks it when using UDP.

This patch change the riemann module to rely on riemann_communicate() to
send events and avoid filling-in the Recv-Q.  Because the function
returns a more granular riemann_message_t, update the trace message to
include these details, and adjust the error detection code accordingly.

Fixes #2521

Signed-off-by: Romain Tartière <romain@blogreen.org>