python-source: fix crash when posting from python thread
Fixes: https://github.com/balabit/syslog-ng/issues/2368

In syslog-ng, logwriters are lazily created: the first message
triggers the creation, and the creation is done by the posting
thread. To avoid race condition, the process is done through a
`main_loop_call()` function. So that to work, on needs to initialize
the posting thread with main_loop_call_thread_init(). However, this
api is not exposed to python, hence posting from thread cannot work
from a python thread other than the one started by the threaded
source.

This patch saves the assigned thread id, and compares the current
thread id of the post_message caller. In case of mismatch, exception
is thrown.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>