test_msgparse: get rid off the determine_year_for_month() dependency
Instead of dynamically calculating timestamps based on the current year,
just fake the current time and make the testing simpler.

The determine_year_for_month() function is tested through the new timeutils
unit tests.

All _get_epoch_with_bsd_year() calls in the tests were replaced as if
the current time was in 2019 and the time is faked accordingly.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>