mainloop: stop worker threads before reverting to the old config
If a new configuration turns out to be invalid during reload, the old one has
to be restored.

In such case, the new configuration is in a half-initialized state, so it
needs to be deinitialized.
This must include a step called main_loop_worker_sync_call(), where worker
threads are requested and waited to be stopped.

Signed-off-by: László Várady <laszlo.varady@balabit.com>