light: separating positional parameters from options
After this patch, the positional parameters and options are stored in
the SourceDriver, DestinationDriver superclasses. After separating the
positionall_parameters and options, the driver codes could be
simplified: they do not need to pass around a made-up parameter name
that was used so that positional parameters could be handled similarly
to options.

As a side effect, positional parameters are not mandatory anymore. For
example example-msg-generator does not have positional parameters.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>