kafka: remove flags(sync)
The implementation of the flags(sync) flag is not really useful, as it will
wait for each and every message to be acknowledged before sending a
subsequent message.

A better solution is to send a batch of messages to keep each partition busy
and wait for acknowledgements as they arrive.  If the delivery of a message
fails, that one can be reposted at the time of the failure.

This patch removes this synchronous implementation, in favour of a future
ack based one.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>