kafka: use key based partitioning
Instead of using the private partitioning implementation, allow the use
of a key for every message. By default the key is not set, but you
can specify it as a template in the key() option of the driver.

If it's set to that string, it will drive the partition selection
(using murmur2_random by default) AND it will also be delivered as a key
to the broker.

This patch also removes the private partitioning we had earlier.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>