.travis: use docker based style validation
This patch will change the travis configuration to use the ubuntu bionic
docker image to do the code style validation. The reason is to avoid having
to use an old version of astyle or maintaining a newer version.

So far we hosted an updated astyle in OBS, that we will not have to do
with this patch.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>