kafka: add support for delivery_reports
This patch enables delivery_reports and makes sure that failed messages
are posted again by pushing them back to the queue.

At this point all such messages go back to the queue of the first worker,
regardless which thread "produced" it. This shouldn't be a problem, as
failed deliveries should be small in numbers, so we are not overloading
a single thread. Also, I am yet to see if it makes sense to run
multiple workers, as the only workload those do is formatting messages
and handing them off to rdkafka.

NOTE: messages that don't get delivered when we exit/reload will be lost
and LEAKED. rdkafka 0.11.5 doesn't have rd_kafka_purge(), but that's coming
in 1.0.0. Without that, we can't really free those messages as we pass
on a reference of the LogMessage associated with each produce.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>