afamqp: add support for external authentication
AMQP has support for an 'external' authentication scheme which uses some
out-of-band method for authenticating the user. For instance, the user
may be identified from the Distinguished Name or Common Name in their
client certificate. The method by which the user is identified when
using external authentication depends on the server configuration.

Add a new 'auth-method' parameter to the amqp destination configuration
which may be set to 'plain' or 'external'. When this parameter is
omitted or set to 'plain', the existing behaviour is unchanged. When set
to 'external' then external authentication is enabled and the username,
if supplied, will be used as the identity, although this is generally
not needed and may be omitted.

Signed-off-by: Chris Spencer <christopher.spencer@sea.co.uk>