cfg: add cfg_is_shutting_down() function
This function allows drivers to determine if we are deinitializing
for a reload or a shutdown. At shutdown we should do a complete flush
whereas during reload we can be less strict as the LogQueue instances
remain intact. At shutdown we only retain the messages in the
queue if the user is using a disk queue.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>