logthrdestdrv: add "expedite" option to flush
The new option makes it possible to know that we are going down for a reload.
If expedite is TRUE, we shouldn't take much time to perform the flush
operation, rather we can simply return LTR_RETRY to indicate
that we want to reprocess the entire batch once the new config
is initialized.

This should speed up reload operations significantly if the target
server is slow.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>