app-parser: fix crash with grammar error
See the following example.

```
@version: 3.20
application a[my-apparser] {
    rewrite { set("b" value("b") };
};

log {
  parser { app-parser(topic(my-apparser)); };
};
```

In this case, the error happens in the application_options in the
grammar file (KW_REWRITE is not allowed), after we freed `$2` and
`$4`. However, in case of grammar error, lexer frees these tokens
too, hence double free.

As a resolution, the `free` calls are moved to the end of the grammar file.
This way only grammar will free these tokens.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>