python: use py2's py_datetime_to_logstamp for py3
After the modification of py2 version of this function,
it is compatible with py3, and as py2 version uses local_timezone,
if *naive* datetime is supplied by the user, it also fixes incompatiblity
issue between py2 and py3.

The only downside of this, that it no longer uses the builtin *timestamp*
method in case of py3.

Signed-off-by: Kokan <kokaipeter@gmail.com>