format-json: refactor type handling to be a bit more easier to read
Earlier the type hinted representation of values was a bit complex,
containing a shared case in a switch statement and then a lot of
conditionals around the key of the switch.

This one breaks that one up, as I was planning to handle the "empty-value"
case here, when I found a better option of doing the same in
the type hinting code.

Nevertheless this version I consider easier to read, so albeit this
could be improved even more, I think this is worth checking in.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>