cmake: removing ENABLE_BASICFUNCS option
This does not work: as there are other test that depend on it, but the
guard is missing around them. In autotools, this module is not
switchable. So we get closer to the autotools version. Basicfuncs has
no external dependency, and we typically make modules modules
switchable, if they have such dependency.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>