logproto-framed-server: extract memmove operations
From now on frame header reading will ensure that there is enough
space in the buffer for reading the frame header. (With a low
chance of unsuccessfull parsing atempt.)
And with the knowledge of the actual frame_len it will also make
sure that later, there will be enough room for the message.

Signed-off-by: Laszlo Szemere <laszlo.szemere@balabit.com>