nondumpable-allocator: fixing mmap error handling
mmap returns -1 instead of NULL in case of error, because in
some (extreme) cases, NULL can be start of a mmapped area. Aligning the
code according to this.

Users of nondumpable-allocator need not be changed. Nondumpable
allocator puts metadata to the front of a page, and returns the memory
area after the metadata. Hence, NULL is never returned after a
successful allocation, so NULL can still be used for error handling.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>