test_diskq: fix possible crash in test
The *test_diskq_statistics* could crash under certain conditions.
If the linux ASLR is enabled, and the compiler creates executable
that is position independent(-fPIC).

The function pointer could became invalid, in the unit test.
As the Criterion does *fork* and *exec* before executing tests.

Additionally the parameters are created in the *parent*,
and passed down to the *child* process, which - because of ASLR -
possible stores the referenced function in a different address.

This patch does not solves the above issue, but hides this behavior
with additionally calling the function pointer passed by
via function pointer.

Because of this, a *@PLT* trampoline function is created, which
address is used, that can be made position independent hence
valid after *fork* plus *exec* if compiled with *-fno-pie*,
which is done in our case.

Signed-off-by: Kokan <kokaipeter@gmail.com>