wildcard-source: merge deleted_file_finished() into deleted_file_eof()
_remove_file_reader() had to be delayed to a point where all acks have
been received in order to avoid persist name collision with a new file
with the same name/path.

Now that bookmark handling is disabled after the file is removed and EOF
is reached, this delay is not necessary anymore.

Pending non-bookmarked acks still keep LogReader and FileReader "alive"
(through LogReader::control), but that instance is deinited and not
relevant for the source driver anymore.

Signed-off-by: László Várady <laszlo.varady@balabit.com>