logsource: add heuristic threshold options
They are percentage values (0-100).

When the average free window is below dynamic_window_increase_threshold,
the window size will be doubled.

When the average free window is above dynamic_window_decrease_threshold,
the window size will be halved (being careful not to go under the initial
window size or the current window).

Signed-off-by: László Várady <laszlo.varady@balabit.com>