filter: Fix "!=" filter
The KW_NUM_NE and KW_NE keywords did not set any comparison flag
for `fop_cmp_eval()`. It was handled by checking for no flag set
(`self->cmp_op == 0`), which is not true in case of KW_NUM_NE,
because with `KW_NUM_NE` the `FCMP_NUM` flag is set.

After this the `|| self->cmp_op == 0` check is not needed,
but we have to make a boolean from `self->cmp_op & FCMP_LT`,
which was passively done by having a `||` after it.
I cleaned it up with double negation.

Signed-off-by: Attila Szakacs <attila.szakacs@balabit.com>