system: remove unused app-parser
App-parsers usage, including using the "system-unix" topic
introduced in 286e1a64b001f12e6be2b4bf5cccab4b21c2081b.

I think removing an unused code is a good step,
especially that it could cause problems when the first system-unix topic
would be introduced.

Reason:
- app-parser creates a channel, which groups some junctions-channels
- 2 app-parsers following each other therefore means 2 channels, i.e. a log pipe sequence.
If a drop happens in the first part, then it will never reach the second part, i.e. the 2nd app-parser topic.

Conclusion is that we either remove the unused app-parser,
or fix it with an additional junction-channel/if-else.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>
Signed-off-by: Antal Nemes <antal.nemes@balabit.com>